Create Prime Factors Of A Number
This Java program finds and prints all the prime factors of a given number. It includes a function `primeFactors(int n)` that takes an integer `n` as input and prints its prime factors. The main function prompts the user to enter a number and then calls the `primeFactors` function to display the results.

## How to Use

1. Clone the repository.
2. Compile and run the Java program.
3. Enter a number when prompted.
4. The program will output the prime factors of the entered number.

Feel free to use, modify, and contribute!